Information about compatibility with AndroidX
We are currently not planning to update this library for support of the Android Support Library version 28 or the new AndroidX libraries, as has been discussed in #62 and other places. Instead, we are trying to get our modifications merged into the official Material Components Android library. Please see the PR for more details.